requirements.txt
Pillow
pyheif
